core: relocate kam reload request services
