+++ library/Ivoz/Kam/Domain/Service/Trusted/SendTrunksPermissionsReloadRequest.php
@@ -1,0 +1,34 @@
+<?php
+
+namespace Ivoz\Kam\Domain\Service\Trusted;
+
+use Ivoz\Kam\Domain\Model\Trusted\TrustedInterface;
+use Ivoz\Kam\Domain\Service\TrunksClientInterface;
+
+class SendTrunksPermissionsReloadRequest implements TrustedLifecycleEventHandlerInterface
+{
+    const ON_COMMIT_PRIORITY = self::PRIORITY_LOW;
+
+    protected $trunksClient;
+
+    public function __construct(
+        TrunksClientInterface $trunksClient
+    ) {
+        $this->trunksClient = $trunksClient;
+    }
+
+    public static function getSubscribedEvents()
+    {
+        return [
+            self::EVENT_ON_COMMIT => self::ON_COMMIT_PRIORITY
+        ];
+    }
+
+    /**
+     * @return void
+     */
+    public function execute(TrustedInterface $trusted)
+    {
+        $this->trunksClient->reloadTrustedPermissions();
+    }
+}
