+++ library/Ivoz/Provider/Domain/Service/Domain/SendUsersDomainReloadRequest.php
@@ -1,0 +1,34 @@
+<?php
+
+namespace Ivoz\Provider\Domain\Service\Domain;
+
+use Ivoz\Kam\Domain\Service\UsersClientInterface;
+use Ivoz\Provider\Domain\Model\Domain\DomainInterface;
+
+class SendUsersDomainReloadRequest implements DomainLifecycleEventHandlerInterface
+{
+    const ON_COMMIT_PRIORITY = self::PRIORITY_NORMAL;
+
+    protected $usersClient;
+
+    public function __construct(
+        UsersClientInterface $usersClient
+    ) {
+        $this->usersClient = $usersClient;
+    }
+
+    public static function getSubscribedEvents()
+    {
+        return [
+            self::EVENT_ON_COMMIT => self::ON_COMMIT_PRIORITY
+        ];
+    }
+
+    /**
+     * @return void
+     */
+    public function execute(DomainInterface $entity)
+    {
+        $this->usersClient->reloadDomain();
+    }
+}
