--- library/Ivoz/Provider/Infrastructure/Domain/Service/ApplicationServer/SendUsersDispatcherReloadRequest.php
@@ -1,35 +1,0 @@
-<?php
-
-namespace Ivoz\Provider\Infrastructure\Domain\Service\ApplicationServer;
-
-use Ivoz\Kam\Domain\Service\UsersClientInterface;
-use Ivoz\Provider\Domain\Model\ApplicationServer\ApplicationServerInterface;
-use Ivoz\Provider\Domain\Service\ApplicationServer\ApplicationServerLifecycleEventHandlerInterface;
-
-class SendUsersDispatcherReloadRequest implements ApplicationServerLifecycleEventHandlerInterface
-{
-    const ON_COMMIT_PRIORITY = self::PRIORITY_HIGH;
-
-    protected $usersGearmanClient;
-
-    public function __construct(
-        UsersClientInterface $usersGearmanClient
-    ) {
-        $this->usersGearmanClient = $usersGearmanClient;
-    }
-
-    public static function getSubscribedEvents()
-    {
-        return [
-            self::EVENT_ON_COMMIT => self::ON_COMMIT_PRIORITY
-        ];
-    }
-
-    /**
-     * @return void
-     */
-    public function execute(ApplicationServerInterface $entity)
-    {
-        $this->usersGearmanClient->reloadDispatcher();
-    }
-}
