--- library/Ivoz/Provider/Infrastructure/Domain/Service/OutgoingRouting/SendTrunksLcrReloadRequest.php
@@ -1,35 +1,0 @@
-<?php
-
-namespace Ivoz\Provider\Infrastructure\Domain\Service\OutgoingRouting;
-
-use Ivoz\Kam\Domain\Service\TrunksClientInterface;
-use Ivoz\Provider\Domain\Model\OutgoingRouting\OutgoingRoutingInterface;
-use Ivoz\Provider\Domain\Service\OutgoingRouting\OutgoingRoutingLifecycleEventHandlerInterface;
-
-class SendTrunksLcrReloadRequest implements OutgoingRoutingLifecycleEventHandlerInterface
-{
-    const ON_COMMIT_PRIORITY = self::PRIORITY_NORMAL;
-
-    protected $trunksClient;
-
-    public function __construct(
-        TrunksClientInterface $trunksClient
-    ) {
-        $this->trunksClient = $trunksClient;
-    }
-
-    public static function getSubscribedEvents()
-    {
-        return [
-            self::EVENT_ON_COMMIT => self::ON_COMMIT_PRIORITY
-        ];
-    }
-
-    /**
-     * @return void
-     */
-    public function execute(OutgoingRoutingInterface $entity)
-    {
-        $this->trunksClient->reloadLcr();
-    }
-}
